leaf82: modularize JS macro proof cases
